+++ Numismatics.CORE/Repositories/IRepository.cs
@@ -8,10 +8,10 @@ namespace Numismatics.CORE.Repositories
 {
     public interface IRepository<T>
     {
-        public T? Get(int id);
+        public T? Get(long id);
         public T? Create(T entity);
         public T? Update(T entity);
-        public T? Delete(int id);
+        public T? Delete(long id);
         public List<T> GetAll();
     }
 }
